// lib/api/recipe/types/API_RecipeByTitle_Response.d.ts
// Generated by https://quicktype.io

export interface API_RecipeByTitle_Response {
    data: Data;
}

interface Data {
    recipeCollection: RecipeCollection;
}

interface RecipeCollection {
    items: RecipeCollectionItem[];
}

interface RecipeCollectionItem {
    sys: Sys;
    tagsCollection: TagsCollection;
    title: string;
    photo: Photo;
    description: string;
    chef: Chef | null;
    calories: number;
}

interface Photo {
    url: string;
    description: string | null;
    title: string;
}

interface Chef {
    name: string;
}

interface Sys {
    id: string;
}

interface TagsCollection {
    items: TagsCollectionItem[];
}

interface TagsCollectionItem {
    name: string;
}